refactor: styled-components 공통부분 모으기
