+++ src/components/addBook/ModalWrapper.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import styled from "styled-components";
+
+import { StBookModalWrapper } from "../common/styled/BookModalWrqpper";
 
 interface ModalWrapperProps {
   children: React.ReactNode;
@@ -7,22 +8,5 @@ interface ModalWrapperProps {
 export default function ModalWrapper(props: ModalWrapperProps) {
   const { children } = props;
 
-  return <StModalWrapper>{children}</StModalWrapper>;
+  return <StBookModalWrapper>{children}</StBookModalWrapper>;
 }
-
-export const StModalWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  width: 100%;
-  height: 100%;
-
-  position: fixed;
-  z-index: 10;
-  left: 0;
-  top: 0;
-
-  background-color: ${({ theme }) => theme.colors.white};
-`;
