+++ src/components/common/styled/Button.tsx
@@ -9,3 +9,24 @@ export const Button = styled.button`
 
   color: ${({ theme }) => theme.colors.white}; // active
 `;
+
+export const StAddAnswerButton = styled.button`
+  width: 6.6rem;
+  height: 3.4rem;
+
+  ${({ theme }) => theme.fonts.caption}
+  color: ${({ theme }) => theme.colors.gray400};
+`;
+
+export const StMenuBtn = styled(Button)`
+  border-radius: 0.8rem;
+  background-color: transparent;
+  width: 9.5rem;
+  height: 3.8rem;
+  ${({ theme }) => theme.fonts.caption}
+  color: ${({ theme }) => theme.colors.gray200};
+
+  &:hover {
+    background-color: ${({ theme }) => theme.colors.white300};
+  }
+`;
