+++ src/components/common/styled/Icon.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+import { IcMore } from "../../../assets/icons";
+
+export const StMoreIcon = styled(IcMore)`
+  margin-right: 1.6rem;
+
+  &:hover {
+    fill: #efefef;
+  }
+`;
