+++ src/components/common/styled/Img.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const StBookCardImg = styled.img`
+  width: 100%;
+  height: 100%;
+
+  /* 뷰 이슈로 곡률 완화 */
+  border-radius: 1.4rem;
+
+  object-fit: cover;
+`;
