+++ src/components/common/styled/Link.tsx
@@ -1,0 +1,11 @@
+import { Link } from "react-router-dom";
+import styled from "styled-components";
+
+export const StLoginLink = styled(Link)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 100%;
+  height: 100%;
+`;
