+++ src/components/common/styled/MenuWrapper.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const StMenuWrapper = styled.div<{ menuposition?: string }>`
+  display: none;
+
+  position: absolute;
+  top: ${({ menuposition }) => (menuposition === "isPriQ" ? "6rem" : menuposition === "isPriA" ? "2.9rem" : "4.3rem")};
+  right: ${({ menuposition }) => (menuposition === "isPriQ" ? "4.4rem" : "1.6rem")};
+  z-index: 10;
+
+  border: 0.1rem solid ${({ theme }) => theme.colors.gray200};
+  border-radius: 0.8rem;
+  padding: 0.4rem;
+  background-color: ${({ theme }) => theme.colors.white};
+`;
