+++ src/components/common/styled/StepModalWrapper.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+import { StBookModalWrapper } from "./BookModalWrqpper";
+
+export const StStepModalWrapper = styled(StBookModalWrapper)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 20;
+
+  background-color: rgba(55, 56, 62, 0.8);
+`;
