+++ src/components/common/styled/Toggle.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+import { IcToggle } from "../../../assets/icons";
+
+export const StIcToggle = styled(IcToggle)`
+  min-width: 2.6rem;
+  margin-left: 1.6rem;
+  &:hover {
+    cursor: pointer;
+  }
+`;
