+++ src/components/detail/ExamplePreNote.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 import theme from "../../styles/theme";
-import LabelQuestion from "../common/styled/LabelQuestion";
+import LabelQuestion from "../common/LabelQuestion";
 
 interface ExamplePreNoteProps {
   answerOne: string | undefined;
