+++ src/components/signup/LastStep.tsx
@@ -4,10 +4,11 @@ import { useNavigate, useOutletContext } from "react-router-dom";
 import styled from "styled-components";
 
 import { ImgSignUpFinish } from "../../assets/images";
-import { StHeading2, UserData } from "../../pages/Signup";
+import { UserData } from "../../pages/Signup";
 import { patchData, postData } from "../../utils/lib/api";
 import { Loading } from "../common";
 import { Button } from "../common/styled/Button";
+import { StSignupHeading2 } from "../common/styled/Signup";
 
 export default function LastStep() {
   const [userData] = useOutletContext<[UserData]>();
@@ -61,11 +62,11 @@ export default function LastStep() {
           initial={{ opacity: 0 }}
           animate={{ opacity: 1 }}
           exit={{ opacity: 0 }}>
-          <StHeading2>
+          <StSignupHeading2>
             {userData["nickname"]}님!
             <br />
             나만의 서재가 완성됐어요!
-          </StHeading2>
+          </StSignupHeading2>
 
           <img src={ImgSignUpFinish} alt="회원 가입 완료시 뜨는 그래픽 문구입니다" />
           <StHomeBtn type="button" onClick={goNextStep}>
